Fix the race in harvest_core's diagnostics::tests::messages.
The test generated tool reporters from separate threads, which allowed the two `tool_a` ToolReporters to be created in a different order from the intended order of the test. This changes the test to construct the tool reporter from the test case's main thread, which solves the race and is more consistent with how harvest_translate's `Runner` uses it.

I also changed the failure output so it dumps the entire contents of the tool run's messages file, which will make test failures easier to debug in the future.